requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
argostranslate==1.9.0
openai==1.3.5
pyyaml==6.0.1
python-dateutil==2.8.2
django-extensions==3.2.3
django-filter==23.3
Pillow==10.1.0
reportlab==4.0.7
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
gunicorn==21.2.0
whitenoise==6.6.0
